seperated out queries and also seperated routes out in to seperate files
