+++ server/plugins/router.js
@@ -1,5 +1,19 @@
 const routes = [].concat(
-  require('../routes/floods')
+  require('../routes/floods'),
+  require('../routes/flood-area-alert'),
+  require('../routes/flood-area-warning'),
+  require('../routes/floods-within'),
+  require('../routes/forcast-data'),
+  require('../routes/impacts-within'),
+  require('../routes/impacts'),
+  require('../routes/is-england'),
+  require('../routes/outlook'),
+  require('../routes/station'),
+  require('../routes/stations-upstream-downstream'),
+  require('../routes/station-telemetry'),
+  require('../routes/stations-within'),
+  require('../routes/stations-within-radius'),
+  require('../routes/thresholds-data')
 )
 
 module.exports = {
